UI update & japanese pois list add
